Changed a little bit the structure of the project: created a "server" folder inside "src"; moved everything inside "src" to "src/server"; moved "src/server/lambdas/fetch-location" to "src/fetch-location-lambda"
