+++ src/fetch-location-lambda/fetch-location-handler.ts
