+++ src/fetch-location-lambda/tsconfig.fetch-location-lambda.json
@@ -1,8 +1,8 @@
 {
-  "extends": "../../../tsconfig.json",
+  "extends": "../../tsconfig.json",
   "compilerOptions": {
-    "outDir": "dist",
     "baseUrl": "./",
+    "outDir": "dist",
   },
   "include": ["fetch-location-handler.ts"],
 }
