+++ src/server/tsconfig.backend-lambda.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "baseUrl": "./",
+    "outDir": "dist",
+  },
+  "include": ["backend-lambda.ts"]
+}
